Implement /isPermutation/ using helper functions
Implemented a helper that creates character counters objects.
Implemented a helper that checks for shallow equality of objects;
will do a strict equality on the values of the objects so will
return false if the values are objects. Works for the purpose of this
problem since objects being checked contain only number values.